+++ src/hooks/index.ts
@@ -16,3 +16,4 @@ export * from './useTransactionManager'
 export * from './useAsset'
 export * from './useDialog'
 export * from './useStorageProviders'
+export * from './useStartFileUpload'
