+++ src/views/studio/EditVideoSheet/EditVideoSheet.tsx
@@ -18,7 +18,6 @@ import {
   useEditVideoSheet,
   useJoystream,
   useTransactionManager,
-  useUploadsManager,
 } from '@/hooks'
 import { CreateVideoMetadata, VideoAssets, VideoId } from '@/joystream-lib'
 import { writeUrlInCache, writeVideoDataInCache } from '@/utils/cachingAssets'
@@ -58,7 +57,7 @@ export const EditVideoSheet: React.FC = () => {
   // transaction management
   const [thumbnailHashPromise, setThumbnailHashPromise] = useState<Promise<string> | null>(null)
   const [videoHashPromise, setVideoHashPromise] = useState<Promise<string> | null>(null)
-  const { startFileUpload } = useUploadsManager()
+  const startFileUpload = useStartFileUpload()
   const { joystream } = useJoystream()
   const { fee, handleTransaction } = useTransactionManager()
   const client = useApolloClient()
